abstract packagemanagers to be portable
